backend update

+++ src/main/java/com/sportcity/demo/entities/Ice_arena.java
@@ -11,7 +11,7 @@ import javax.persistence.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
-public class Ice_arena extends AbstractEntity<Long>{
+public class Ice_arena extends AbstractEntity<Integer>{
 
     @Column(name = "square")
     private Integer square;
