+++ src/main/java/com/sportcity/demo/repositories/OrganizerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizerRepository extends JpaRepository<Organizer, Long> {
+public interface OrganizerRepository extends JpaRepository<Organizer, Integer> {
 }
