+++ src/main/java/com/sportcity/demo/repositories/SportsmanRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SportsmanRepository extends JpaRepository<Sportsman, Long> {
+public interface SportsmanRepository extends JpaRepository<Sportsman, Integer> {
 }
